+++ pyspelling/__main__.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 import sys
 import argparse
 from pyspelling import settings
-from pyspelling import Aspell
+from pyspelling import spellcheck
 from pyspelling import __version__
 
 
@@ -15,14 +15,16 @@ def run():
     parser.add_argument('--version', action='version', version=('%(prog)s ' + __version__.version))
     parser.add_argument('--verbose', '-v', action='count', default=0, help="Verbosity level.")
     parser.add_argument('--name', '-n', action='store', default='', help="Specific spelling task by name to run.")
-    parser.add_argument('--binary', '-b', action='store', default='', help="Provide path to Aspell binary.")
+    parser.add_argument('--binary', '-b', action='store', default='', help="Provide path to spell checker's binary.")
     parser.add_argument('--config', '-c', action='store', default='.spelling.yml', help="Spelling config.")
+    parser.add_argument(
+        '--spellchecker', '-s', action='store', default='', help="Choose between aspell and hunspell"
+    )
     args = parser.parse_args()
 
     fail = False
     config = settings.read_config(args.config)
-    spelling = Aspell(config, name=args.name, binary=args.binary, verbose=args.verbose)
-    for results in spelling.check():
+    for results in spellcheck(config, args.name, args.binary, args.verbose, args.spellchecker):
         if results.error:
             fail = True
             print('ERROR: Could not read %s -- %s' % (results.context, results.error))